crag_rag/requirements.txt
torch
sentence-transformers
faiss-cpu
transformers
scikit-learn
numpy 